Frontend/requirements.txt
streamlit
pandas
numpy
joblib
scikit-learn
xgboost
shap
matplotlib
seaborn
mlxtend
firebase-admin
plotly
networkx
python-dotenv